+++ pkg/config/schema_validators.go
@@ -143,201 +143,6 @@ const (
 	maxWireI32 = int64(math.MaxInt32)
 )
 
-// ValidateBGPHoldTime accepts a BGP hold-time in seconds: 0, or 3..65535.
-// FRR (and RFC 4271 §4.2 / §10) require the hold-time to be either 0
-// (hold timer disabled) or at least 3 seconds — the keepalive interval is
-// hold-time/3, so a hold-time of 1 or 2 yields a sub-second/zero keepalive
-// that FRR's `neighbor X timers <keepalive> <hold>` command REJECTS, and a
-// rejected line fails the WHOLE frr-reload (a single `vtysh -f` add-batch
-// exits non-zero on any CMD_WARNING_CONFIG_FAILED) — one bad leaf takes
-// down every routing change on that reload. The renderer treats 0 as "unset"
-// via its `> 0` gate (no `timers` line), an accepted documented tradeoff, so
-// 0 is permitted here; 1 and 2 are the values that must be rejected at commit
-// before they can reach frr-reload. Upper bound is the 16-bit on-wire max.
-func ValidateBGPHoldTime(raw string, _ *Config) error {
-	if strings.TrimSpace(raw) == "" {
-		return fmt.Errorf("missing value (expected integer)")
-	}
-	v, err := strconv.ParseInt(raw, 10, 64)
-	if err != nil {
-		return fmt.Errorf("not an integer: %q", raw)
-	}
-	if v != 0 && (v < 3 || v > 65535) {
-		return fmt.Errorf("BGP hold-time must be 0 or 3..65535 seconds (got %d); FRR rejects 1/2 and a rejected timers line fails the whole reload", v)
-	}
-	return nil
-}
-
-// routeFilterMatchTypes are the match-type keywords accepted in a
-// `from route-filter <prefix> <match-type>` node. The schema node is
-// args:2, so the walker invokes ValidateRouteFilterArg for BOTH identity
-// tokens (the prefix AND the match-type — Keys[1:3]); the validator must
-// therefore accept a match-type keyword as well as a CIDR prefix.
-//
-// exact/longer/orlonger/upto/prefix-length-range are the rendered
-// match-types (pkg/frr policy_render.go). `through` is ADMITTED as a
-// grammar token here but REJECTED at commit by
-// validateRouteFilterMatchTypesStrict (#2525): FRR prefix-lists express
-// only length ranges (ge/le) and cannot represent the two-prefix
-// containment path of Junos `through`. It stays in this set so the
-// commit-check reaches the semantic gate (which emits an actionable
-// "unsupported match-type" error) rather than failing earlier with a
-// generic "not a valid prefix" message. `prefix-length-range` is
-// admitted AND rendered (`ge low le high`); its /low-/high bounds are
-// semantically validated by validateRouteFilterMatchTypesStrict.
-var routeFilterMatchTypes = map[string]bool{
-	"exact":               true,
-	"longer":              true,
-	"orlonger":            true,
-	"upto":                true,
-	"prefix-length-range": true,
-	"through":             true,
-}
-
-// ValidateRouteFilterArg is the #2105 commit-check validator for a
-// `policy-options policy-statement <p> term <t> from route-filter
-// <prefix> <match-type>` identity arg token. The node is args:2, so the
-// walker (schema_walk.go) calls this for the prefix token AND the
-// match-type token. The `upto /N` length token lands as a CHILD of the
-// route-filter node, never in Keys[1:3], so it does not reach here.
-//
-// A token is accepted when it is either a match-type keyword OR a
-// syntactically valid CIDR prefix (v4 or v6 — route-filter is
-// family-agnostic). A malformed prefix (no "/", non-numeric mask, or an
-// out-of-range mask) that is neither is REJECTED, so it never reaches
-// the FRR renderer where it would produce an FRR-invalid prefix-list
-// line and could fail the whole managed frr-reload.
-//
-// Strictness is automatic: SchemaValidate is strict on the operator
-// commit / commit-check path and lenient (warn, not fail) on Store.Load
-// / SyncApply (the #1960 lenient-downgrade-on-load doctrine), so a
-// prefix persisted by an older binary never blackouts boot. The renderer
-// carries a belt-and-suspenders skip for any malformed prefix that
-// reaches it via that lenient path.
-//
-// Known limitation: the validator is position-agnostic (the walker does
-// not tell it which slot a token came from), so a match-type keyword in
-// the prefix slot (e.g. `route-filter longer exact`) is accepted. This
-// is strictly better than the pre-#2105 state (any garbage committed)
-// and catches the real malformed-CIDR case; per-position validation
-// would require schema restructuring and is deferred.
-func ValidateRouteFilterArg(raw string, _ *Config) error {
-	tok := strings.TrimSpace(raw)
-	if tok == "" {
-		return fmt.Errorf("missing route-filter prefix (expected CIDR, e.g. 10.0.0.0/24)")
-	}
-	if routeFilterMatchTypes[tok] {
-		return nil
-	}
-	// Family-agnostic CIDR. parseCIDRStrict requires a /prefix-length
-	// and upgrades the common operator mistakes (bare IP, garbage) to
-	// targeted messages, matching the ValidateIPv4CIDR/ValidateIPv6CIDR
-	// convention; the returned IP is discarded because both families are
-	// valid here. Surface parseCIDRStrict's targeted message (e.g.
-	// "missing /prefix-length") so the commit-check failure is actionable.
-	if _, err := parseCIDRStrict(tok, "10.0.0.0/24"); err != nil {
-		return fmt.Errorf("not a valid route-filter prefix (expected a CIDR, e.g. 10.0.0.0/24 or 2001:db8::/32, or a match-type keyword): %v", err)
-	}
-	return nil
-}
-
-// ValidateRouteDestination accepts a static-route destination prefix: a
-// family-agnostic CIDR (v4 or v6) with an explicit /prefix-length. Used for
-// the `routing-options static route <destination>` identity arg (#2448):
-// the destination feeds net.ParseCIDR in the FRR renderer
-// (pkg/frr/config_render.go generateStaticRoute) and net/ipnet parse in the
-// Rust FIB builder (userspace-dp forwarding_build/fib.rs populate_routes).
-// A destination that parses as neither IPv4 nor IPv6 is SILENTLY dropped by
-// both consumers today — the route commits cleanly but never installs. The
-// default routes 0.0.0.0/0 and ::/0 parse via net.ParseCIDR and are
-// accepted; a bare IP without a length, or outright garbage, is rejected
-// with a targeted message. route is family-agnostic so both families pass.
-func ValidateRouteDestination(raw string, _ *Config) error {
-	// parseCIDRStrict requires a /prefix-length and upgrades the two common
-	// operator mistakes (bare IP, garbage) to targeted messages. The returned
-	// IP is discarded because both v4 and v6 destinations are valid here.
-	if _, err := parseCIDRStrict(raw, "10.0.0.0/24"); err != nil {
-		return fmt.Errorf("not a valid route destination (expected a CIDR, e.g. 10.0.0.0/24 or 2001:db8::/32): %v", err)
-	}
-	return nil
-}
-
-// ValidateStaticNextHop accepts a static-route next-hop VALUE: a bare IPv4
-// or IPv6 address, the Rust-FIB `ip@interface` / `@interface` spec, or a
-// bare interface name. It rejects values that are genuinely malformed —
-// neither a usable IP nor a plausible interface name — so an operator typo
-// fails loud at commit instead of installing a blackhole (#2448).
-//
-// Why each accepted form is real:
-//   - a bare IP (192.168.1.1, 2001:db8::1) is the common gateway form; the
-//     FRR renderer emits it verbatim (config_render.go) and the Rust FIB
-//     parses it (parse_route_next_hop[_v6]).
-//   - `ip@iface` / `@iface` is the Rust FIB interface-scoped form
-//     (forwarding_build/fib.rs split_once('@')); the IP part, when present,
-//     must itself parse, else the spec silently degrades to interface-only.
-//   - a bare interface name (ge-0-0-0.0, reth0.50, eth1) is a valid Junos
-//     next-hop and renders as an interface route in FRR (config_render.go
-//     `case ifName != ""`).
-//
-// The malformed cases rejected: a botched IP literal (1.2.3.999,
-// 2001:db8::garbage), an `ip@iface` whose IP part does not parse
-// (notanip@eth0 — silently becomes interface-only), and any value that is
-// neither a valid IP nor a syntactically plausible interface name. The
-// interface-name test requires at least one ASCII letter so a numeric-only
-// dotted value (a botched IPv4) cannot masquerade as an interface, and uses
-// the [A-Za-z0-9._-] charset interface names actually use — which excludes
-// ':' so a botched IPv6 literal is rejected rather than mistaken for a name.
-func ValidateStaticNextHop(raw string, _ *Config) error {
-	tok := strings.TrimSpace(raw)
-	if tok == "" {
-		return fmt.Errorf("missing next-hop (expected an IP address, e.g. 192.168.1.1 or 2001:db8::1, or an interface name)")
-	}
-	// Rust-FIB ip@interface / @interface spec.
-	if ipPart, ifPart, hasAt := strings.Cut(tok, "@"); hasAt {
-		if ifPart == "" {
-			return fmt.Errorf("malformed next-hop %q: missing interface name after '@'", raw)
-		}
-		if ipPart != "" && net.ParseIP(ipPart) == nil {
-			return fmt.Errorf("malformed next-hop %q: %q is not a valid IP address before '@interface'", raw, ipPart)
-		}
-		if !plausibleInterfaceName(ifPart) {
-			return fmt.Errorf("malformed next-hop %q: %q is not a valid interface name after '@'", raw, ifPart)
-		}
-		return nil
-	}
-	if net.ParseIP(tok) != nil {
-		return nil
-	}
-	if plausibleInterfaceName(tok) {
-		return nil
-	}
-	return fmt.Errorf("not a valid next-hop (got %q; expected an IP address, ip@interface, or an interface name)", raw)
-}
-
-// plausibleInterfaceName reports whether s is syntactically usable as a
-// next-hop interface name: the [A-Za-z0-9._-] charset that xpf/vSRX and
-// kernel interface names use, with at least one ASCII letter so a
-// numeric-only dotted token (a botched IPv4 like 1.2.3.999) cannot pass as
-// an interface name. ':' is excluded so a malformed IPv6 literal is not
-// mistaken for an interface name.
-func plausibleInterfaceName(s string) bool {
-	if s == "" {
-		return false
-	}
-	hasLetter := false
-	for _, r := range s {
-		switch {
-		case r >= 'a' && r <= 'z', r >= 'A' && r <= 'Z':
-			hasLetter = true
-		case r >= '0' && r <= '9', r == '.', r == '-', r == '_':
-			// allowed non-letter
-		default:
-			return false
-		}
-	}
-	return hasLetter
-}
-
 // ValidateDHGroup accepts a Diffie-Hellman group as either a bare integer
 // (e.g. "14") or the Junos "group<N>" spelling (e.g. "group14") — both are
 // configured in the wild and both compile (compiler_ipsec.go strips the
